requirements.txt
Flask
requests
numpy
gunicorn
google-generativeai
python-dotenv
torch
sentence-transformers
